Add lot to StockMeta

+++ app/src/storage/stocks/stocksstorage.cpp
@@ -529,7 +529,8 @@ static void getTurnoverForParallel(
 
             for (int i = index; i < stock->data.size(); ++i)
             {
-                totalTurnover += qRound64(stock->data.at(i).quantity * stock->data.at(i).price);
+                totalTurnover +=
+                    qRound64(static_cast<double>(stock->data.at(i).quantity) * static_cast<double>(stock->data.at(i).price));
             }
 
             const qint64 deltaTimestamp = stock->data.constLast().timestamp - stock->data.at(index).timestamp;
